upd:string:string_insertAt

+++ libs/Labs/Headers/lab1.h
@@ -6,7 +6,7 @@
 #define LAB1_H
 
 String* decimalConverter(int num,char (*func)(int*));
-String* floatConverter(double num,int system,int precision);
+String* floatConverter(double num,size_t system,int precision);
 
 char asHex(int* num);
 char asBinary(int *num);
